v0.6.2 UPDATE
- Introduced DMFT_CTRL_VARS: a module to collect and store control variables.
The scope of the module is to create a dictionary or memory pool where control variables, e.g.
temperature, local interaction, chemical potential, are saved and can be retrieved on request.
This is especially useful in procedure that partially depend on user defined control variables
that can not be passed directly as argument of the procedure itself.
So one now can set some control variable in the calling program, that will be searched for in an
internal procedure inside DMFT_TOOLS (e.g. dmft_get_kinetic_energy, that will search for beta,
inverse temperature).

We are now ready to detach modules from different projects and create:
.DMFT_KINETIC_ENERGY
.DMFT_GLOC
.DMFT_WEISS_FIELD