v0.6.4 NEW:
Introduced DMFT_WEISS_FIELD module.
- This module performs the calculations to get the local Weiss Field
or Hybridization function \Delta for the normal and superconducting case,
both for a single site or Nlat ineq. sites.

Fixed bugs in:
- DMFT_EKIN: there was a small bug in the subroutine for the lattice normal case.
The Lmats variable was erroneously extracted from the lenght of the 3rd
dimension of the input Sigma array was. It was supposed to be the 4th.

Updates:
- DMFT_GLOC: the iprint flag in the lattice case now runs from 0 to 6 (default),
1,2,3 print in compressed mode (No Split), 4,5,6 print in splitted mode (one plot per site).

Tests for DMFT_EKIN and DMFT_GLOC are passed in the normal case.
Testing the DMFT_WEISS_FIELD module.