v0.7.2 UPDATE:
- updated the DMFT_EKIN module for kinetic energy calculation in DMFT.
The structure of the module has been simplified using included files.
Subroutines are used now in place of functions, optional Ekin and Eloc
variables are included as input for each routine.
The Ekin and Eloc variables are now array of size(Hk,1) (Nso or Nlso).
The routine evaluate the contributions to kinetic and local energy
for each spin-orbital or site-spin-orbital.
Total and partial values are then printed to file.
For back-compatibility the total values are still in first place (first
two values of the first row).

Routine have been tested just for
1b Bethe
1b and 2b Attractive Square