New Subs
Added and tested nightmarish subroutine that computes the eigenbands starting from
the Wannier90 output H(R). Since, depending on the K point on the path,
the computation could be demanding, the sub does as many things as possible
once the Fourier transform is done. Even the correction with Re{impSigma(w=0)}.
This is why there are so many optional arguments.
All the subs related to w90 are for spin-conserving systems (no SOC).