v1.0.8 BUG FIX
We fixed a bug in DMFT_EKIN.

F.Grandi found that a term Hloc was missing in the definition of the
Green's function in all the _dos procedure (those evaluating the
kinetic energy for systems defined thru their DOS).
This has been added for all cases (normal/superc serial/mpi).

Wait to confirm bug is fixed.