1.2.7 Update
Merged with fix_TB_w90

This version introduces a better handling of the W90 interface.
This is achieved as follow.

A structure is defined: w90_structure, containing all the informations
relevant for a W90 file.

Then we divided the work in two main parts:
- setup (TB_w90_setup / TB_w90_delete)
- model (TB_w90_model >> TB_build_model, TB_solve_model)

The setup part performs the file reading and setup,allocate,define
all the necessary arrays, e.g.
- Number of wannier orbitals,
- Number of Wigner-Seitz vectors,
- Read multiplicity or degeneracies of the vectors,
- Read vectors & tight binding hoppings amplitudes
- Create the real-space vector grid
- Check direct lattice basis vectors have been defined
- Check the that dimensions of the problems are correct

For, this procedure needs some information such as:
+ w90 file
+ Nlat = Number of inequivalent sites in the lattice unit cell
+ Norb = Number of orbitals per ineq. site
+ Nspin= Spin degeneracy (1=non-magnetic, 2=magnetic)

The structure can be deallocated with TB_w90_delete.

Once setup is performed, the system automatically create an
abstract function TB_w90_model, which transforms the collected
information into a model system. TB_w90_model takes as input
a k-point and Hamiltonian leading dimension N. Outputs Hk
as a matrix Hk(N,N) which corresponds to the Hamiltonian for that
k-point.

This abstract function performs a single Fourier transforms from
real-space to k-space.
TB_w90_model can be used in the calling code in place of any
hk_model function input in TB_{build,solve,...}_model. So that
many of the drivers should not be changed.

As a second method the procedures TB_{build,solve,...} have been
extended so to include usage with Wannier 90 files. Just the input
hk_model has been removed from the list of input.

A simple test is provided in the project TBmodels/hm_2d_w90.
The 3 (2x W90 + standard Hk model) methods have been tested for the square lattice using the
following hand-written Wannier 90 file:

written <some date>
 1
 4
 1 1 1 1
 1    0   0    1    1   -1.000000   -0.000000
 0    1   0    1    1   -1.000000   -0.000000
-1    0   0    1    1   -1.000000   -0.000000
 0   -1   0    1    1   -1.000000   -0.000000

which corresponds to a square lattice with hopping t=-1.

The results obtained in the three cases are identical.
More tests to come for a more serious band structure.

TODO: extend to have MPI support, i.e. have all nodes
reading the Wannier 90 file (bcasting the structure is way
harder).