1.2.9 Update + bug fix
Small updates and additions to Wannier90 interface in TB.

The new procedure TB_w90_FermiLevel evaluates (for a given
grid of k-points) the Fermi level at a given filling (per spin).
The corresponding w90_model function is then shifted automatically
at the calculated Fermi level. This avoid problems when the bands are
high in energy, i.e. the Fermi level is far away.

A similar function is created for the model part.

TODO: add pushing of the Z function and ReSigma(0) to shift and renormalize
the W90 model, this is useful to retrieve and plot renormalized bands.

FIx a bug in the reshape of the H(k) matrix from w90 model (call to
slo2lso error)