1.2.13 UPDATE:
The TB_solve_model alogn path has been updated again.
- Updates extended to the H(r,k) hybrid form case.
- Simplified plotting script using the GNUPLOT every statement to
plot all blocks or a single one (corresponding to one or more
lattice-spin-orbitals). A comment explains how to use it, just in case.
- Added iproject flag (optional) to actually  project the k-path onto
the lattice basis. This MUST be used if the path is defined in absolute
terms and not with respect to the basis, then a projection is required.
Default: iproject = TB_w90%status
so if W90 is used projection is automatic and must be disabled explicitly.