Added the possibility to handle bethe multiorbital calculations in the presence of local off-diagonal matrix elements (hybrid/replica bath).
The modification in DMFT_GLOC consist in a new interface procedure in the aforementioned case.
It is motivated by the fact that we need to perform a matrix inversion rather than the inversion of each diagonal element separately.
Then, since I want to preserve the possibility of having different DoS (Dbands) for different orbitals, the code provides as output only diagonal entries.
This is because its not clear to me which DoS should I use to integrate off-diagonal Gfs.

The modification in DMFT_WEISS_FIELD consist in the modification of the existing procedure that now
for the Bethe self-consitency is computing  Delta = T*G*T in the dumb way.

All this bla bla compiles but, since I don't want to bother other users (and I don't have time now to
include the same modifications in all the other cases, i.e. serial and superc), I am pushing everything on a different branch.