1.2.14 UPDATE
Major update in TIGHT_BINDING module.
The main module has been exploded into several smaller modules.
MPI support has been directly included in the build up of the
model (both model + wannier90) this speeds up a lot the execution
when dealing with many k-points.
The MPI supports is based on the use of pre-processing directives
ifdef _MPI
and the use of check_MPI() function (just introduced in Scifor)
to check initilization of MPI so to be able to use or not the main
communicator MPI_COMM_WORLD in an automatic way: i.e.
if user init MPI then this is used to construct H(k) and other
functions, if not we fall back to scalar.

TODO: this kind of setup will be extended to the other
modules where MPI is heavily used (DMFT_GLOC,WEISS,EKIN,etc.)