1.2.16 NEW FEATURE
Fermi Surface calculations in TIGHT_BINDING.
+ Added new procedures in TB_COMMON
+ Added simple Grid refining scheme in TB_BASIS
+ Implemented a primitive but effective Fermi Surface calculation using
grid refining.
The logic is to execute a three-step passage (arbitrary step algorithm
need a specific data structure). Anyhow already 3 iteration can take
very long time if the initial grid is very dense.
1. Create a regular grid of the BZ and look for FS points.
2. For each of the N points just found, create a new finer grid
of width 4*dx,4*dy centered at the point. Look for FS point
3. Iterate one more time using the same scheme .
At each step the cutoff_ is reduced by a factor 10.
It is important to start from a large value, e.g. 0.1-0.3
but not too high as to include too many initial points.

On output a file + GNUplot script is produced.
Orbital colors are required on input.
Tested on 1bands and 2bands model on square lattice.

With 1-step procedure a grid as fine as 500*500 was required to resolve
the FS.
With a 3-step procedure a grid of 50*50 is already more than enough to
get essentially exact results.

To visualize the BZ shifted in -pi:pi the logical flag pi_shift=T/F can be used

More tests to come. Enjoy.