Added feature in dmft_get_gloc_matsubara_normal_dos for non diagonal Smats
The subroutine now can compute Gloc_mats for non diagonal Smats with bands all equal up to a shift and a reshape.
Note that for this new case size(Dbands,1)==1 and must be explicitly implemented in the driver.
The shift is contained in Hloc. The reshape must be such that the passed Dbands has Wband=1.0
The old code, lines 305-331, runs if dos_diag=.true.
The new code, lines 334-353, runs if dos_diag=.false.