Added feature in dmft_get_gloc_realaxis_normal_dos for non-diag Sreal
The subroutine now can compute Gloc_real for non diagonal Sreal with bands all equal up to a shift and a reshape.
Note that for this new case size(Dbands,1)==1 and must be explicitly implemented in the driver.
The shift is contained in Hloc. The reshape must be such that the passed Dbands has Wband=1.0
The old code runs if dos_diag=.true.
The new code runs if dos_diag=.false.