Added feature in DoS Ekin computation
This new version can compute Ekin and Eloc for non-diagonal Smats with bands all equal up to a shift and a reshape.
Note that for this new case size(Dbands,1)==1 and must be explicitly implemented in the driver.
The shift is contained in Hloc. The reshape must be such that the passed Dbands has Wband=1.0

The previous version runs if dos_diag=.true.
The new part runs if dos_diag=.false.