2.1.3 Update in DMFT_TIGHT_BINDING
The library has been updated with new features and corrections to the
Wannier90 interface.
1. We introduced a TB_reorder==TB_reshape functions, limited to
vectors and matrices (array rank 1 and 2). These functions aim to
re-order the input arrays according to a user specified form in terms
of Nlat, Norb, Nspin variables.
*Problem*: the arrays are usually built according to some order, e.g.
[Norb<Nspin<Nlat] the convention used for DMFT calculations (where
different lattice site have to be treated separately).
However this convention or a given order is sometime not valid, leading to
wrong results. For instance W90 interface oftens build H(Ri,Rj)-H(k) with the
order [Norb<Nlat<Nspin]. In general W90 results can have an arbitrary order
of the elements which is decided in the configuration file (not accessed in out
interface).

This requires one to re-order the array such to make them compatible with the
calculation method to be used.
Solution:TB_reorder==TB_reshape accomplish this task by reorganizing the array.
For, the old and the new ordering should be specified at input as:
Huser   = the array (rank1-2, dble-cmplx)
Nin     = a rank-1, dimension(3) integer array with the input values of Nlat,Norb,Nspin
OrderIn = a rank-1, dimension(3) character(len=5) array with the INPUT order specified by the
strings "Nlat", "Nspin", "Norb" from the innermost (faster) to the outermost (slower)
OrderOut= OUTPUT order, same type and rules of INPUT.

* W90 Interface. setup_w90 and w90_structure.
We updated the W90 data structure and the corresponding setup function to
be adherent with the re-ordering issue.
-Nspin is now an optional,integer argument (default 1)
-Spinor is a new logical,optional argument indicating if Wannier90 calculation was done with spinor flag
True or Fals, i.e. if spin degrees of freedom are included in wannier orbitals.
Note that Spinor=T => Nspin=2, but Nspin=2 can be with Spinor=F generating a two-spin-blocks Hamiltonian
whic MUST be reordered.
> THE REORDERING IS NOT AUTOMATIC <
- Header is a new logical,optional argument indicating is W90 file contains or not the header information.

TESTING TO BE DONE.