2.2.3 UPDATE: convergence check for 4-dimensional arrays (3 + frequency)
These routines were part of the extension of the code that addressed
Real space + CDMFT. That series of commits created problems with
compilers older than fortran 8, for which rank 8 arrays are not
supported, and has been addressed via precompilation flags.

Since in the DMFT_CONVERGENCE folder some routines are split into .f90
files to be included, the use of precompilation flags is not so simple
in this case, and these functions have been discarded in the merge.

However, though the convergence functions are indeed used in the real space
+ CDMFT case,  they never deal with 8+ rank arrays, therefore their
inclusion should be safe in general.