2.4.0 Update + New features
+ TB_build_kgrid received an additional, optional, input variable: width
dble,dimension(:)
This variable express the extension in the [0:1] interval of the kgrid to be generated.
This can be used in situations were you DO NOT want to sample the entire BZ but just a
subset of it. Note that you can re-center the Kgrid using :origin as input variable still
in the [0:1] intervals, because every point is generated in the hypercube [0:1]^d and then
mapped onto the BZ using the reciprocal lattice basis.

+ TB_build_kgrid includes and additional interface to :refine_kgrid with the following:
interface
subroutine TB_build_kgrid(kgridIN,Nkvec,kcenters,lambda,kgrid)
    real(8),dimension(:,:)                :: kgridIN
    integer,dimension(:)                  :: Nkvec
    real(8),dimension(:,:)                :: kcenters
    real(8),dimension(size(Nkvec))        :: lambda
    real(8),dimension(:,:),allocatable    :: kgrid
end subroutine
end interface

kgridIN is a given Kgrid in the BZ, generated using the same TB_build_kgrid([Nx,Ny,Nz],kgrid)
Nkvec is a vector of integers expressing the grid size along each direction for the refining.
kcenters is a list of points shape(Ncenters,Ndim) which contains the point around which the
grid must get thicker
lambda is the width of the thicker region AROUND each kcenter (-lambda/2:lamdba/2) along each
direction
kgrid is the OUTPUT kgrid and it is allocated and built inside the routine.

e.g. a 10x10 grid in [0:2pi]x[0:2pi] and its refinement with no overlap around [0,0] using
20x20 grid. Note that only 1/4 of the grid survives as 3/4 of it extend outside BZ and the
corresponding points are discarded.

  +------------------------------------------------------------------------+
  |                                                                        |
  |       A       A       A       A        A       A       A       A       |
  |                                                                        |
  |       A       A       A       A        A       A       A       A       |
  |                                                                        |
  |                                                                        |
  |       A       A       A       A        A       A       A       A       |
  |                                                                        |
  |       A       A       A       A        A       A       A       A       |
  |                                                                        |
  |       A       A       A       A        A       A       A       A       |
  |                                                                        |
  |       A       A       A       A        A       A       A       A       |
  |                                                                        |
  | A A A A A A A A A                                                      |
  | A A A A A A A A A     A       A        A       A       A       A       |
  | A A A A A A A A A                                                      |
  | A A A A A A A A A     A       A        A       A       A       A       |
  | A A A A A A A A A                                                      |
  +------------------------------------------------------------------------+