2.4.2 Minor bug fix + update
+ The Grid refine system introduced in 2.4.0 and updated in 2.4.1 was
missing a precious information about the number of points.

The algorithm now search&delete any point in the IN-Kgrid which overlaps
with the refining grids. The remaining points are included in the OUT-Kgrid
and counted (NkClean), this number is saved into the first NkClean component
of an weight array.
Then refining grids are generated with NkRefined points, added to the weight
array and the OUT-Kgrid.

Update takes care of periodicity of the BZ so that search&delete algorithm
now automatically wraps the refining grids aroung [0:1) (before mapping to BZ
using basis lattice vector is performed).