Intermediate commit.
I restored the old module DMFT_GF to evaluate local and k-resolved GF in
matsubara e realaxis. This is done to fulfill back compatibility but it
should be considered as a temporary solution.
> Note however that GF writing has been updated so calls to these procedures
should be updated in the drivers:
call dmft_write_gf(Func,"File",axis="{mats,real}",iprint=1-5,...)

TEST:
single and multi band GF has been evaluated with the old and new methods.
Results are identical within the diff.
results for real-space lattice have error of 1d-16 or smaller.

Ready to merge.