2.6.1 New Feature
Included the procedure:
TB_w90_transform(func)

in TB_WANNIER90. This can be used to implement some specific transformation
onto the real-space Hamiltonian H_ab(R=[Rx,Ry,Rz]), with a,b=1,...,NumWann.

The procedure accepts a function as input. The function is defined on the user side
to act on the Hamiltonian matrix.
The function *func can be of two type (interfaces)
+func(H,N), complex*8 H(N,N), integer N
The Hamiltonian H_ab for each one R=1,...,Nrpts is passed to H(N,N)

+func(H,N,Nr) complex*8 H(N,N,Nr), integer N, integer Nr
The Hamiltonian H_ab(R) (all vector R) is passed to H(N,N,Nr)

>NOTE: This procedure accomplish a different task from TB_reshape, whose
goal is to re-order the degrees of freedom in rank-1,2 V(:),M(:,:) and rank-3 Hk(:,:,1:Nk)