Intermediate commit:
set explicit flag "diagonal" in GLOC from DOS calculation routine to
explicitly force the assumption DOS is not diagonal, even if it is. The
self-energy might not be. I am not too sure of what the previous
discriminant actually did, but I left it as default when no explicit
request is made